// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  user: 'user',
  subAdmin: 'subAdmin',
  admin: 'admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Gender = {
  male: 'male',
  female: 'female'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const Department = {
  architecture: 'architecture',
  chemicalEngineering: 'chemicalEngineering',
  civilEngineering: 'civilEngineering',
  electricalAndComputerEngineering: 'electricalAndComputerEngineering',
  electromechanicalEngineering: 'electromechanicalEngineering',
  environmentalEngineering: 'environmentalEngineering',
  mechanicalEngineering: 'mechanicalEngineering',
  miningEngineering: 'miningEngineering',
  softwareEngineering: 'softwareEngineering',
  biotechnology: 'biotechnology',
  foodScienceAndAppliedNutrition: 'foodScienceAndAppliedNutrition',
  geology: 'geology',
  industrialChemistry: 'industrialChemistry',
  postgraduate: 'postgraduate',
  other: 'other'
} as const

export type Department = (typeof Department)[keyof typeof Department]